ledger-on-sql: Use PostgreSQL's `RETURNING` keyword to save a query.
