+++ ledger/ledger-on-sql/src/main/scala/com/daml/ledger/on/sql/queries/H2Queries.scala
@@ -8,6 +8,8 @@ import java.sql.Connection
 import anorm.SqlParser._
 import anorm._
 import com.daml.ledger.on.sql.queries.Queries.Index
+import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlLogEntryId
+import com.google.protobuf.ByteString
 
 class H2Queries extends Queries with CommonQueries {
   override def createLogTable()(implicit connection: Connection): Unit = {
@@ -22,9 +24,15 @@ class H2Queries extends Queries with CommonQueries {
     ()
   }
 
-  override def lastLogInsertId()(implicit connection: Connection): Index =
+  override def insertIntoLog(
+      entry: DamlLogEntryId,
+      envelope: ByteString,
+  )(implicit connection: Connection): Index = {
+    SQL"INSERT INTO log (entry_id, envelope) VALUES (${entry.getEntryId.toByteArray}, ${envelope.toByteArray})"
+      .executeInsert()
     SQL"CALL IDENTITY()"
       .as(long("IDENTITY()").single)
+  }
 
   override protected val updateStateQuery: String =
     "MERGE INTO state VALUES ({key}, {value})"
