+++ ledger/ledger-on-sql/src/main/scala/com/daml/ledger/on/sql/queries/Queries.scala
@@ -26,9 +26,7 @@ trait Queries {
   def insertIntoLog(
       entry: DamlKvutils.DamlLogEntryId,
       envelope: ByteString,
-  )(implicit connection: Connection): Unit
-
-  def lastLogInsertId()(implicit connection: Connection): Index
+  )(implicit connection: Connection): Index
 
   def selectStateByKeys(
       keys: Iterable[DamlKvutils.DamlStateKey],
